chore(expr): move random_column() from Schema to Column
