+++ src/query/expression/src/schema.rs
@@ -12,57 +12,28 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::borrow::Cow;
 use std::collections::BTreeMap;
 use std::collections::BTreeSet;
 use std::collections::HashMap;
 use std::collections::HashSet;
 use std::sync::Arc;
 
-use common_arrow::arrow::bitmap::Bitmap;
 use common_arrow::arrow::datatypes::DataType as ArrowDataType;
 use common_arrow::arrow::datatypes::Field as ArrowField;
 use common_arrow::arrow::datatypes::Schema as ArrowSchema;
 use common_arrow::arrow::datatypes::TimeUnit;
 use common_exception::ErrorCode;
 use common_exception::Result;
-use ethnum::i256;
 use itertools::Itertools;
-use jsonb::Number as JsonbNumber;
-use jsonb::Object as JsonbObject;
-use jsonb::Value as JsonbValue;
-use rand::distributions::Alphanumeric;
-use rand::distributions::DistString;
-use rand::rngs::SmallRng;
-use rand::Rng;
-use rand::SeedableRng;
 use serde::Deserialize;
 use serde::Serialize;
 
-use crate::types::array::ArrayColumn;
-use crate::types::date::DATE_MAX;
-use crate::types::date::DATE_MIN;
-use crate::types::decimal::DecimalColumn;
 use crate::types::decimal::DecimalDataType;
 use crate::types::decimal::DecimalSize;
-use crate::types::nullable::NullableColumn;
-use crate::types::timestamp::TIMESTAMP_MAX;
-use crate::types::timestamp::TIMESTAMP_MIN;
-use crate::types::BooleanType;
 use crate::types::DataType;
-use crate::types::DateType;
 use crate::types::NumberDataType;
-use crate::types::NumberType;
-use crate::types::StringType;
-use crate::types::TimestampType;
-use crate::types::VariantType;
-use crate::with_number_mapped_type;
 use crate::with_number_type;
-use crate::BlockEntry;
-use crate::Column;
-use crate::FromData;
 use crate::Scalar;
-use crate::Value;
 use crate::ARROW_EXT_TYPE_EMPTY_ARRAY;
 use crate::ARROW_EXT_TYPE_EMPTY_MAP;
 use crate::ARROW_EXT_TYPE_VARIANT;
@@ -1023,187 +994,6 @@ impl TableDataType {
             TableDataType::String => "VARCHAR".to_string(),
             TableDataType::Nullable(inner_ty) => format!("{} NULL", inner_ty.sql_name()),
             _ => self.to_string().to_uppercase(),
-        }
-    }
-
-    pub fn create_random_column(&self, len: usize) -> BlockEntry {
-        match self {
-            TableDataType::Null => BlockEntry {
-                data_type: DataType::Null,
-                value: Value::Column(Column::Null { len }),
-            },
-            TableDataType::EmptyArray => BlockEntry {
-                data_type: DataType::EmptyArray,
-                value: Value::Column(Column::EmptyArray { len }),
-            },
-            TableDataType::EmptyMap => BlockEntry {
-                data_type: DataType::EmptyMap,
-                value: Value::Column(Column::EmptyMap { len }),
-            },
-            TableDataType::Boolean => BlockEntry {
-                data_type: DataType::Boolean,
-                value: Value::Column(BooleanType::from_data(
-                    (0..len).map(|_| SmallRng::from_entropy().gen_bool(0.5)),
-                )),
-            },
-            TableDataType::String => BlockEntry {
-                data_type: DataType::String,
-                value: Value::Column(StringType::from_data((0..len).map(|_| {
-                    let rng = SmallRng::from_entropy();
-                    rng.sample_iter(&Alphanumeric)
-                        // randomly generate 5 characters.
-                        .take(5)
-                        .map(u8::from)
-                        .collect::<Vec<_>>()
-                }))),
-            },
-            TableDataType::Number(num_ty) => BlockEntry {
-                data_type: DataType::Number(*num_ty),
-                value: Value::Column(with_number_mapped_type!(|NUM_TYPE| match num_ty {
-                    NumberDataType::NUM_TYPE => NumberType::<NUM_TYPE>::from_data(
-                        (0..len).map(|_| SmallRng::from_entropy().gen())
-                    ),
-                })),
-            },
-            // useless for now.
-            TableDataType::Decimal(t) => match t {
-                DecimalDataType::Decimal128(x) => BlockEntry {
-                    data_type: DataType::Decimal(*t),
-                    value: Value::Column(Column::Decimal(DecimalColumn::Decimal128(
-                        vec![0i128; len].into(),
-                        *x,
-                    ))),
-                },
-                DecimalDataType::Decimal256(x) => BlockEntry {
-                    data_type: DataType::Decimal(*t),
-                    value: Value::Column(Column::Decimal(DecimalColumn::Decimal256(
-                        vec![i256::ZERO; len].into(),
-                        *x,
-                    ))),
-                },
-            },
-            TableDataType::Timestamp => BlockEntry {
-                data_type: DataType::Timestamp,
-                value: Value::Column(TimestampType::from_data(
-                    (0..len)
-                        .map(|_| SmallRng::from_entropy().gen_range(TIMESTAMP_MIN..=TIMESTAMP_MAX))
-                        .collect::<Vec<i64>>(),
-                )),
-            },
-            TableDataType::Date => BlockEntry {
-                data_type: DataType::Date,
-                value: Value::Column(DateType::from_data(
-                    (0..len)
-                        .map(|_| SmallRng::from_entropy().gen_range(DATE_MIN..=DATE_MAX))
-                        .collect::<Vec<i32>>(),
-                )),
-            },
-            TableDataType::Nullable(inner_ty) => {
-                let entry = inner_ty.create_random_column(len);
-                BlockEntry {
-                    data_type: DataType::Nullable(Box::new(entry.data_type)),
-                    value: Value::Column(Column::Nullable(Box::new(NullableColumn {
-                        column: entry.value.into_column().unwrap(),
-                        validity: Bitmap::from(
-                            (0..len)
-                                .map(|_| SmallRng::from_entropy().gen_bool(0.5))
-                                .collect::<Vec<bool>>(),
-                        ),
-                    }))),
-                }
-            }
-            TableDataType::Array(inner_ty) => {
-                let mut inner_len = 0;
-                let mut offsets: Vec<u64> = Vec::with_capacity(len + 1);
-                offsets.push(inner_len);
-                for _ in 0..len {
-                    inner_len += SmallRng::from_entropy().gen_range(0..=3);
-                    offsets.push(inner_len);
-                }
-                let entry = inner_ty.create_random_column(inner_len as usize);
-                BlockEntry {
-                    data_type: DataType::Array(Box::new(entry.data_type)),
-                    value: Value::Column(Column::Array(Box::new(ArrayColumn {
-                        values: entry.value.into_column().unwrap(),
-                        offsets: offsets.into(),
-                    }))),
-                }
-            }
-            TableDataType::Map(inner_ty) => {
-                let mut inner_len = 0;
-                let mut offsets: Vec<u64> = Vec::with_capacity(len + 1);
-                offsets.push(inner_len);
-                for _ in 0..len {
-                    inner_len += SmallRng::from_entropy().gen_range(0..=3);
-                    offsets.push(inner_len);
-                }
-                let entry = inner_ty.create_random_column(inner_len as usize);
-                BlockEntry {
-                    data_type: DataType::Map(Box::new(entry.data_type)),
-                    value: Value::Column(Column::Map(Box::new(ArrayColumn {
-                        values: entry.value.into_column().unwrap(),
-                        offsets: offsets.into(),
-                    }))),
-                }
-            }
-            TableDataType::Tuple { fields_type, .. } => {
-                let mut fields = Vec::with_capacity(len);
-                let mut types = Vec::with_capacity(len);
-                for field_type in fields_type.iter() {
-                    let entry = field_type.create_random_column(len);
-                    fields.push(entry.value.into_column().unwrap());
-                    types.push(entry.data_type);
-                }
-                BlockEntry {
-                    data_type: DataType::Tuple(types),
-                    value: Value::Column(Column::Tuple(fields)),
-                }
-            }
-            TableDataType::Variant => {
-                let mut data = Vec::with_capacity(len);
-                for _ in 0..len {
-                    let opt = SmallRng::from_entropy().gen_range(0..=6);
-                    let val = match opt {
-                        0 => JsonbValue::Null,
-                        1 => JsonbValue::Bool(true),
-                        2 => JsonbValue::Bool(false),
-                        3 => {
-                            let s = Alphanumeric.sample_string(&mut rand::thread_rng(), 5);
-                            JsonbValue::String(Cow::from(s))
-                        }
-                        4 => {
-                            let num = SmallRng::from_entropy().gen_range(i64::MIN..=i64::MAX);
-                            JsonbValue::Number(JsonbNumber::Int64(num))
-                        }
-                        5 => {
-                            let arr_len = SmallRng::from_entropy().gen_range(0..=5);
-                            let mut values = Vec::with_capacity(arr_len);
-                            for _ in 0..arr_len {
-                                let num = SmallRng::from_entropy().gen_range(i64::MIN..=i64::MAX);
-                                values.push(JsonbValue::Number(JsonbNumber::Int64(num)))
-                            }
-                            JsonbValue::Array(values)
-                        }
-                        6 => {
-                            let obj_len = SmallRng::from_entropy().gen_range(0..=5);
-                            let mut obj = JsonbObject::new();
-                            for _ in 0..obj_len {
-                                let k = Alphanumeric.sample_string(&mut rand::thread_rng(), 5);
-                                let num = SmallRng::from_entropy().gen_range(i64::MIN..=i64::MAX);
-                                let v = JsonbValue::Number(JsonbNumber::Int64(num));
-                                obj.insert(k, v);
-                            }
-                            JsonbValue::Object(obj)
-                        }
-                        _ => JsonbValue::Null,
-                    };
-                    data.push(val.to_vec());
-                }
-                BlockEntry {
-                    data_type: DataType::Variant,
-                    value: Value::Column(VariantType::from_data(data)),
-                }
-            }
         }
     }
 }
