+++ src/query/expression/src/values.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::borrow::Cow;
 use std::cmp::Ordering;
 use std::hash::Hash;
 use std::io::Read;
@@ -44,6 +45,8 @@ use crate::property::Domain;
 use crate::types::array::ArrayColumn;
 use crate::types::array::ArrayColumnBuilder;
 use crate::types::boolean::BooleanDomain;
+use crate::types::date::DATE_MAX;
+use crate::types::date::DATE_MIN;
 use crate::types::decimal::Decimal;
 use crate::types::decimal::DecimalColumn;
 use crate::types::decimal::DecimalColumnBuilder;
@@ -63,6 +66,8 @@ use crate::types::string::StringColumn;
 use crate::types::string::StringColumnBuilder;
 use crate::types::string::StringDomain;
 use crate::types::timestamp::check_timestamp;
+use crate::types::timestamp::TIMESTAMP_MAX;
+use crate::types::timestamp::TIMESTAMP_MIN;
 use crate::types::variant::JSONB_NULL;
 use crate::types::*;
 use crate::utils::arrow::append_bitmap;
@@ -71,6 +76,7 @@ use crate::utils::arrow::buffer_into_mut;
 use crate::utils::arrow::constant_bitmap;
 use crate::utils::arrow::deserialize_column;
 use crate::utils::arrow::serialize_column;
+use crate::utils::FromData;
 use crate::with_decimal_mapped_type;
 use crate::with_decimal_type;
 use crate::with_number_mapped_type;
@@ -1481,6 +1487,146 @@ impl Column {
         }
     }
 
+    pub fn random(ty: &DataType, len: usize) -> Self {
+        use jsonb::Number as JsonbNumber;
+        use jsonb::Object as JsonbObject;
+        use jsonb::Value as JsonbValue;
+        use rand::distributions::Alphanumeric;
+        use rand::distributions::DistString;
+        use rand::rngs::SmallRng;
+        use rand::Rng;
+        use rand::SeedableRng;
+
+        // Migrate from legacy code:
+        match ty {
+            DataType::Null => Column::Null { len },
+            DataType::EmptyArray => Column::EmptyArray { len },
+            DataType::EmptyMap => Column::EmptyMap { len },
+            DataType::Boolean => {
+                BooleanType::from_data((0..len).map(|_| SmallRng::from_entropy().gen_bool(0.5)))
+            }
+            DataType::String => StringType::from_data((0..len).map(|_| {
+                let rng = SmallRng::from_entropy();
+                rng.sample_iter(&Alphanumeric)
+                    // randomly generate 5 characters.
+                    .take(5)
+                    .map(u8::from)
+                    .collect::<Vec<_>>()
+            })),
+            DataType::Number(num_ty) => {
+                with_number_mapped_type!(|NUM_TYPE| match num_ty {
+                    NumberDataType::NUM_TYPE => {
+                        NumberType::<NUM_TYPE>::from_data(
+                            (0..len).map(|_| SmallRng::from_entropy().gen()),
+                        )
+                    }
+                })
+            }
+            DataType::Decimal(t) => match t {
+                DecimalDataType::Decimal128(x) => {
+                    Column::Decimal(DecimalColumn::Decimal128(vec![0i128; len].into(), *x))
+                }
+                DecimalDataType::Decimal256(x) => {
+                    Column::Decimal(DecimalColumn::Decimal256(vec![i256::ZERO; len].into(), *x))
+                }
+            },
+            DataType::Timestamp => TimestampType::from_data(
+                (0..len)
+                    .map(|_| SmallRng::from_entropy().gen_range(TIMESTAMP_MIN..=TIMESTAMP_MAX))
+                    .collect::<Vec<i64>>(),
+            ),
+            DataType::Date => DateType::from_data(
+                (0..len)
+                    .map(|_| SmallRng::from_entropy().gen_range(DATE_MIN..=DATE_MAX))
+                    .collect::<Vec<i32>>(),
+            ),
+            DataType::Nullable(ty) => Column::Nullable(Box::new(NullableColumn {
+                column: Column::random(ty, len),
+                validity: Bitmap::from(
+                    (0..len)
+                        .map(|_| SmallRng::from_entropy().gen_bool(0.5))
+                        .collect::<Vec<bool>>(),
+                ),
+            })),
+            DataType::Array(inner_ty) => {
+                let mut inner_len = 0;
+                let mut offsets: Vec<u64> = Vec::with_capacity(len + 1);
+                offsets.push(0);
+                for _ in 0..len {
+                    inner_len += SmallRng::from_entropy().gen_range(0..=3);
+                    offsets.push(inner_len);
+                }
+                Column::Array(Box::new(ArrayColumn {
+                    values: Column::random(inner_ty, inner_len as usize),
+                    offsets: offsets.into(),
+                }))
+            }
+            DataType::Map(inner_ty) => {
+                let mut inner_len = 0;
+                let mut offsets: Vec<u64> = Vec::with_capacity(len + 1);
+                offsets.push(0);
+                for _ in 0..len {
+                    inner_len += SmallRng::from_entropy().gen_range(0..=3);
+                    offsets.push(inner_len);
+                }
+                Column::Map(Box::new(ArrayColumn {
+                    values: Column::random(inner_ty, inner_len as usize),
+                    offsets: offsets.into(),
+                }))
+            }
+            DataType::Tuple(fields) => {
+                let fields = fields
+                    .iter()
+                    .map(|ty| Column::random(ty, len))
+                    .collect::<Vec<_>>();
+                Column::Tuple(fields)
+            }
+            DataType::Variant => {
+                let mut data = Vec::with_capacity(len);
+                for _ in 0..len {
+                    let opt = SmallRng::from_entropy().gen_range(0..=6);
+                    let val = match opt {
+                        0 => JsonbValue::Null,
+                        1 => JsonbValue::Bool(true),
+                        2 => JsonbValue::Bool(false),
+                        3 => {
+                            let s = Alphanumeric.sample_string(&mut rand::thread_rng(), 5);
+                            JsonbValue::String(Cow::from(s))
+                        }
+                        4 => {
+                            let num = SmallRng::from_entropy().gen_range(i64::MIN..=i64::MAX);
+                            JsonbValue::Number(JsonbNumber::Int64(num))
+                        }
+                        5 => {
+                            let arr_len = SmallRng::from_entropy().gen_range(0..=5);
+                            let mut values = Vec::with_capacity(arr_len);
+                            for _ in 0..arr_len {
+                                let num = SmallRng::from_entropy().gen_range(i64::MIN..=i64::MAX);
+                                values.push(JsonbValue::Number(JsonbNumber::Int64(num)))
+                            }
+                            JsonbValue::Array(values)
+                        }
+                        6 => {
+                            let obj_len = SmallRng::from_entropy().gen_range(0..=5);
+                            let mut obj = JsonbObject::new();
+                            for _ in 0..obj_len {
+                                let k = Alphanumeric.sample_string(&mut rand::thread_rng(), 5);
+                                let num = SmallRng::from_entropy().gen_range(i64::MIN..=i64::MAX);
+                                let v = JsonbValue::Number(JsonbNumber::Int64(num));
+                                obj.insert(k, v);
+                            }
+                            JsonbValue::Object(obj)
+                        }
+                        _ => JsonbValue::Null,
+                    };
+                    data.push(val.to_vec());
+                }
+                VariantType::from_data(data)
+            }
+            DataType::Generic(_) => unreachable!(),
+        }
+    }
+
     pub fn remove_nullable(&self) -> Self {
         match self {
             Column::Nullable(inner) => inner.column.clone(),
@@ -1531,7 +1677,7 @@ impl Column {
         }
     }
 
-    /// Returns (is_all_null,  Option bitmap)
+    /// Returns (is_all_null, Option bitmap)
     pub fn validity(&self) -> (bool, Option<&Bitmap>) {
         match self {
             Column::Null { .. } => (true, None),
